final_files/Code/Deployment/requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coverage==5.5
cycler==0.10.0
Cython==0.29.22
dataframe-image==0.1.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.7
django-after-response==0.2.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.11.2
entrypoints==0.3
gast==0.4.0
google-pasta==0.2.0
grpcio==1.36.0
h5py==2.10.0
idna==2.10
importlib-metadata==2.0.0
ipykernel==5.3.4
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
mkl-fft
mkl-random==1.1.1
mkl-service==2.3.0
multidict==5.1.0
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.19.2
olefile==0.46
opt-einsum==3.3.0
packaging==20.9
padelpy==0.1.8
pandas==1.2.2
pandocfilters==1.4.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pip==21.0.1
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.14.0
psycopg2==2.8.6
ptyprocess==0.7.0
PubChemPy==1.0.4
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
scikit-learn==0.23.2
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
setuptools==52.0.0.post20210125
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tb-nightly==1.15.0a20190806
tensorboard==1.14.0
tensorflow==2.0.0rc1
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
tf-estimator-nightly==1.14.0.dev2019080601
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
